Add dynamic confirmation page (#21)
* Work in progress on dynamic cart page

* Add dynamic cart page and add some documentation + refactor some code

* Update cart and cartService scripts

* Add bootstrap 5 validation classes

* Work in progress on form validation

* Add inputs validation

Validate the form's inputs and display errors if needed. Also disable the submit button if the form is not validated successfully.

Bugfix in scss to solve a Bootstrap bug that prevents the cursor to be changed on disabled buttons.

* Modify address regex validation

An address should start nor end with a letter or a digit. An address can contains letters, digits, whitespaces (only one between words), comma, quotes, underscore, hyphen, apostrophe.

* Modify some regex

* Add error messages when the user submit the form without fill any field

* Validate and order products from the cart page

* Add order confirmation page

* Clear cart after payment

* Remove default payment form values

* Display split orders message only if the cart contains more than 1 product type

* Email address validation fix